@@ -67,52 +69,52 @@
     "testURL": "http://localhost/"
   },
   "dependencies": {
-    "@angular/animations": "9.1.11",
-    "@angular/common": "9.1.11",
-    "@angular/compiler": "9.1.11",
-    "@angular/core": "9.1.11",
-    "@angular/forms": "9.1.11",
-    "@angular/localize": "9.1.11",
-    "@angular/platform-browser": "9.1.11",
-    "@angular/platform-browser-dynamic": "9.1.11",
-    "@angular/router": "9.1.11",
-    "@auth0/angular-jwt": "4.2.0",
-    "@ng-bootstrap/ng-bootstrap": "6.1.0",
-    "@swimlane/ngx-datatable": "17.0.0",
+    "@angular/animations": "10.0.4",
+    "@angular/common": "10.0.4",
+    "@angular/compiler": "10.0.4",
+    "@angular/core": "10.0.4",
+    "@angular/forms": "10.0.4",
+    "@angular/localize": "10.0.4",
+    "@angular/platform-browser": "10.0.4",
+    "@angular/platform-browser-dynamic": "10.0.4",
+    "@angular/router": "10.0.4",
+    "@auth0/angular-jwt": "5.0.1",
+    "@ng-bootstrap/ng-bootstrap": "7.0.0",
+    "@swimlane/ngx-datatable": "17.1.0",
     "@types/file-saver": "2.0.1",
     "angular-tree-component": "8.5.6",
-    "async-mutex": "0.2.2",
+    "async-mutex": "0.2.4",
     "bootstrap": "4.5.0",
     "chart.js": "2.8.0",
-    "detect-browser": "5.1.0",
+    "detect-browser": "5.1.1",
     "file-saver": "2.0.2",
     "fork-awesome": "1.1.7",
     "lodash": "4.17.19",
-    "moment": "2.26.0",
-    "ng-block-ui": "2.1.8",
+    "moment": "2.27.0",
+    "ng-block-ui": "3.0.2",
     "ng-bootstrap-form-validation": "5.0.0",
     "ng-click-outside": "6.0.0",
     "ng2-charts": "2.3.0",
-    "ngx-toastr": "12.1.0",
-    "rxjs": "6.5.5",
-    "simplebar-angular": "2.2.0",
-    "swagger-ui-dist": "3.26.0",
-    "tslib": "1.12.0",
+    "ngx-toastr": "13.0.0",
+    "rxjs": "6.6.0",
+    "simplebar-angular": "2.2.1",
+    "swagger-ui-dist": "3.28.0",
+    "tslib": "2.0.0",
     "zone.js": "0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "0.901.7",
-    "@angular/cli": "9.1.7",
-    "@angular/compiler-cli": "9.1.11",
-    "@angular/language-service": "9.1.11",
+    "@angular-devkit/build-angular": "0.1000.3",
+    "@angular/cli": "10.0.3",
+    "@angular/compiler-cli": "10.0.4",
+    "@angular/language-service": "10.0.4",
     "@compodoc/compodoc": "1.1.11",
     "@locl/cli": "0.0.1-beta.9",
-    "@types/jest": "26.0.3",
+    "@types/jest": "26.0.4",
     "@types/lodash": "4.14.157",
-    "@types/node": "12.12.47",
+    "@types/node": "12.12.50",
     "@types/simplebar": "5.1.1",
-    "codelyzer": "5.2.2",
-    "cypress": "4.10.0",
+    "codelyzer": "6.0.0",
+    "cypress": "4.11.0",
     "html-linter": "1.1.1",
     "htmllint-cli": "0.0.7",
     "jest": "26.1.0",
@@ -120,19 +122,19 @@
     "jest-preset-angular": "8.2.1",
     "jest-silent-reporter": "0.2.1",
     "ng-bullet": "1.0.3",
-    "ng-mocks": "10.0.0",
+    "ng-mocks": "10.0.2",
     "npm-force-resolutions": "0.0.3",
     "npm-run-all": "4.1.5",
     "prettier": "2.0.5",
     "pretty-quick": "2.0.1",
-    "start-server-and-test": "1.11.0",
-    "stylelint": "13.6.0",
+    "start-server-and-test": "1.11.2",
+    "stylelint": "13.6.1",
     "stylelint-config-sass-guidelines": "7.0.0",
     "stylelint-declaration-use-variable": "1.7.2",
     "transifex-i18ntool": "1.1.0",
     "ts-node": "8.10.2",
     "tslint": "6.1.2",
-    "typescript": "3.8.3"
+    "typescript": "3.9.6"
   },
   "resolutions": {
     "fsevents": "2.1.2",
